clarify introductory component example (#526)
* Update index.md

I tried to follow the guide and every example worked, except for the components. Upon further investigation, I found the issue: the component must be registered before the Vue instance. This was not clear in the Introductory guide, so I suggest it's added ( I copied the same sentence as seen on https://vuejs.org/guide/components.html )

* clarify introductory component example